requirements.txt
streamlit
langchain
langchain_groq
langchain_core
python-dotenv
langchain_community
duckduckgo-search
yfinance
google-search-results